--- src/main/kotlin/domain/Card.kt
@@ -1,7 +1,0 @@
-package domain
-
-class Card(val symbol: Symbol, val value: Value) {
-    fun name(): String {
-        return value.valueName + symbol.symbol
-    }
-}
