+++ src/main/kotlin/domain/gamer/Player.kt
@@ -1,4 +1,4 @@
-package domain
+package domain.gamer
 
 class Player(name: String, hand: Hand = Hand()) : Gamer(name, hand) {
 }
