+++ src/main/kotlin/domain/gamer/Players.kt
@@ -1,4 +1,6 @@
-package domain
+package domain.gamer
+
+import domain.card.Deck
 
 class Players(private val players: List<Player>) : List<Player> by players {
     fun initStage(deck: Deck) {
