+++ src/test/kotlin/domain/gamer/HandTest.kt
@@ -1,5 +1,9 @@
-package domain
+package domain.gamer
 
+import domain.card.Card
+import domain.card.Symbol
+import domain.card.Value
+import domain.gamer.Hand
 import org.assertj.core.api.Assertions
 import org.assertj.core.api.AssertionsForInterfaceTypes.assertThat
 import org.junit.jupiter.api.DisplayName
