+++ src/test/kotlin/domain/gamer/PlayerTest.kt
@@ -1,5 +1,9 @@
-package domain
+package domain.gamer
 
+import domain.card.Card
+import domain.card.Symbol
+import domain.card.Value
+import domain.gamer.Player
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.DisplayName
 import org.junit.jupiter.api.Test
